requirements.txt
asgiref
django-ckeditor
django-js-asset
django-phonenumber-field
django-widget-tweaks
phonenumbers
Pillow
pip
psycopg2
pytz
setuptools
sqlparse
tzdata
wheel